control-plane/pyproject.toml
[project]
name = "control-plane"
version = "0.1.0"
description = "EdgeOrchestra CLI control plane"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.0",
    "rich>=13.9.0",
    "httpx>=0.28.0",
    "grpcio>=1.68.0",
    "protobuf>=5.29.0",
    "zeroconf>=0.136.0",
]

[project.scripts]
eo = "control_plane.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/control_plane"]
